Move visit types into nph order directory
