+++ src/Nph/Order/Visits/VisitDSMT.php
@@ -1,28 +1,30 @@
 <?php
 
-namespace App\Biobank;
+namespace App\Nph\Order\Visits;
 
-class VisitMMTT
+use App\Biobank\Samples;
+
+class VisitDSMT
 {
     public static $timePoints = [
-        'preMMTT' => 'Pre MMTT',
+        'preDSTT' => 'Pre DSTT',
         'minus15min' => '-15 Min',
-        'minus5min' => '-5 Min',
+        'min5min' => '-5 Min',
         '15min' => '15 Min',
         '30min' => '30 Min',
         '60min' => '60 Min',
         '90min' => '90 Min',
         '120min' => '120 Min',
         '240min' => '240 Min',
-        'postMMTT' => 'Post MMTT'
+        'postDSTT' => 'Post DSTT'
     ];
 
-    public static function getTimePointsWithSamples($module): array
+    public function getSamples($module): array
     {
         $module = 'module' . $module;
         $timePointSamples = [];
         foreach (self::$timePoints as $key => $timePoint) {
-            if ($key === 'preMMTT' || $key === 'postMMTT') {
+            if ($key === 'preDSTT' || $key === 'postDSTT') {
                 $timePointSamples[$key] = Samples::${$module . ucfirst($key)};
             } else {
                 $timePointSamples[$key] = Samples::${$module . 'BloodSamples'};
